fix runtime error

+++ src/main/java/ca/uwaterloo/cs451/a1/PairsPMI.java
@@ -65,7 +65,7 @@ public class PairsPMI extends Configured implements Tool {
         public void reduce(PairOfStrings key, Iterable<IntWritable> values, Context context)
             throws IOException, InterruptedException {
             int sum = 0;
-            Iterator<FloatWritable> iter = values.iterator();
+            Iterator<IntWritable> iter = values.iterator();
             while (iter.hasNext()) {
             sum += iter.next().get();
             }
